LayoutTests/imported/w3c/web-platform-tests/web-animations/responsive/columnCount-expected.txt


FAIL column-count responsive to inherited column-count changes assert_equals: expected "3" but got "auto"
